refactor(admin): fold custom /admin page into /admin/browse
The custom React /admin page and its JSON `/admin/collections|tables`
backend predated the axum-admin sketch at /admin/browse. They covered
overlapping ground — both browsed Postgres tables read-only — but
duplicated the table allowlist (the point that triggered #475).

Removes the React surface and JSON API in favor of axum-admin
exclusively:
- 7 frontend files deleted (~625L): AdminPage, CollectionDetailPage,
  TableDetailPage and their stories, services/admin.ts.
- 3 SPA routes removed from App.tsx.
- routes/admin.rs deleted (~213L); AdminAuth moves into
  routes/admin_browse.rs (its only consumer).
- observing-db/src/admin.rs deleted (~340L) — KNOWN_COLLECTIONS,
  KnownTable struct, lookup/count/stats/list_records helpers were
  all only consumed by the JSON routes that just went away.
- The (schema, name) registration list is inlined as ADMIN_TABLES
  in admin_browse.rs. axum-admin pulls column metadata from
  information_schema at request time, so per-table column configs
  are unnecessary; the search_path covers the schema lookups.
- Server-side `GET /admin` 308-redirects to `/admin/browse` so old
  bookmarks keep working.

What's lost: per-NSID stats (count, unique DIDs, oldest/newest
indexed_at), DID-filter UI, cascades-to display, and MUI styling on
admin pages. None of those were load-bearing operationally — psql
+ the row browser cover the same ground. axum-admin's strength is
that it stays small; the lexicon-aware features didn't fit cleanly
inside it and aren't worth the duplication anywhere else.

Supersedes #475 — that PR added the same lexicon record tables to
the now-removed KNOWN_TABLES allowlist; equivalent registrations
live inline in admin_browse.rs here.
