+++ crates/observing-appview/src/routes/admin_browse.rs
@@ -1,38 +1,107 @@
-//! HTML admin browser. Mounts `axum-admin` at `/admin/browse`, gated by the
-//! same DID allowlist as the JSON `/admin/*` endpoints.
+//! HTML admin browser. Mounts `axum-admin` at `/admin/browse`, gated by an
+//! `AdminAuth` extractor that requires a logged-in DID in `ADMIN_DIDS`.
 //!
-//! Tables exposed are pulled from `observing_db::admin::KNOWN_TABLES` so the
-//! HTML browser and the JSON API share a single allowlist.
+//! The list of browsable tables is the source of truth for the admin
+//! surface — there's no parallel JSON API or React page anymore. To
+//! expose another table here, add a `(schema, name)` entry to
+//! `ADMIN_TABLES`. axum-admin pulls column lists from
+//! `information_schema.columns` at request time, so no per-table column
+//! config is needed; the schema must be on the connection's
+//! `search_path` (it is — `ingester, appview, public` is set
+//! database-wide by migration 20260419000000).
+//!
+//! `oauth_state`/`oauth_sessions` are intentionally excluded — browsing
+//! tokens through an admin UI is the failure mode being avoided.
 
 use axum::extract::{FromRequestParts, Request, State};
+use axum::http::request::Parts;
 use axum::middleware::Next;
-use axum::response::{IntoResponse, Response};
+use axum::response::{IntoResponse, Redirect, Response};
 use axum::Router;
-use observing_db::admin::KNOWN_TABLES;
 
-use crate::routes::admin::AdminAuth;
+use crate::auth::AuthUser;
+use crate::error::AppError;
 use crate::state::AppState;
 
-/// Build the `/admin/browse` router with the existing AdminAuth gate applied
-/// as a middleware. Returns a fully-stated `Router` so the caller can
+/// `(schema, name)` pairs registered with axum-admin. Browse-everything
+/// approach: any table here is visible at `/admin/browse/{schema}/{name}`.
+const ADMIN_TABLES: &[(&str, &str)] = &[
+    // Internal state.
+    ("ingester", "ingester_state"),
+    ("ingester", "notifications"),
+    ("ingester", "community_ids"),
+    ("appview", "occurrence_private_data"),
+    ("appview", "notification_reads"),
+    ("public", "sensitive_species"),
+    // Lexicon record tables (occurrences, identifications, comments,
+    // interactions, likes). The browser shows raw rows; per-NSID
+    // slicing is gone with the old custom /admin page.
+    ("ingester", "occurrences"),
+    ("ingester", "identifications"),
+    ("ingester", "comments"),
+    ("ingester", "interactions"),
+    ("ingester", "likes"),
+    // Caches and operational tables.
+    ("ingester", "taxa"),
+    ("ingester", "failed_records"),
+];
+
+/// Build the `/admin/browse` router with the AdminAuth gate applied as a
+/// middleware. Returns a fully-stated `Router` so the caller can
 /// `nest("/admin/browse", router(state))` it under the main app.
 pub fn router(state: AppState) -> Router {
     let mut admin = axum_admin::Admin::new(state.pool.clone());
-    for meta in KNOWN_TABLES {
-        admin = admin.table(meta.schema, meta.name, |t| t);
+    for (schema, name) in ADMIN_TABLES {
+        admin = admin.table(*schema, *name, |t| t);
     }
     admin
         .into_router("/admin/browse")
         .layer(axum::middleware::from_fn_with_state(state, require_admin))
 }
 
-/// Run the `AdminAuth` extractor on every request before passing it through.
-/// Same gate as the JSON admin routes — empty allowlist → 503, missing
-/// session → 401, wrong DID → 403.
+/// Server-side redirect from the legacy `/admin` route to the HTML
+/// browser. The custom React admin page used to live at `/admin`; this
+/// keeps old bookmarks working without dragging a SPA route along.
+pub async fn redirect_to_browse() -> Redirect {
+    Redirect::permanent("/admin/browse")
+}
+
+/// Run the `AdminAuth` extractor on every request before passing it
+/// through. Empty allowlist → 503, missing session → 401, wrong DID → 403.
 async fn require_admin(State(state): State<AppState>, request: Request, next: Next) -> Response {
     let (mut parts, body) = request.into_parts();
     match AdminAuth::from_request_parts(&mut parts, &state).await {
         Ok(_) => next.run(Request::from_parts(parts, body)).await,
         Err(err) => err.into_response(),
+    }
+}
+
+/// Axum extractor that requires a logged-in user whose DID is in
+/// `ADMIN_DIDS`. Rejects with 503 when the allowlist is empty so the
+/// admin surface is opt-in by default.
+pub struct AdminAuth {
+    #[allow(dead_code)]
+    pub did: String,
+}
+
+impl FromRequestParts<AppState> for AdminAuth {
+    type Rejection = AppError;
+
+    async fn from_request_parts(
+        parts: &mut Parts,
+        state: &AppState,
+    ) -> Result<Self, Self::Rejection> {
+        if state.admin_dids.is_empty() {
+            return Err(AppError::ServiceUnavailable(
+                "Admin interface is disabled (ADMIN_DIDS not set)".into(),
+            ));
+        }
+        let user = AuthUser::from_request_parts(parts, state).await?;
+        if !state.admin_dids.iter().any(|d| d == &user.did) {
+            return Err(AppError::Forbidden(
+                "Your account is not authorized for admin access".into(),
+            ));
+        }
+        Ok(AdminAuth { did: user.did })
     }
 }
