+++ crates/observing-db/src/lib.rs
@@ -1,4 +1,3 @@
-pub mod admin;
 pub mod comments;
 pub mod community_ids;
 pub mod failed_records;
