+++ frontend/src/App.tsx
@@ -26,9 +26,6 @@ import { LexiconView } from "./components/lexicon/LexiconView";
 import { DocsPage } from "./components/docs/DocsPage";
 import { NotificationsPage } from "./components/notifications/NotificationsPage";
 import { SettingsPage } from "./components/settings/SettingsPage";
-import { AdminPage } from "./components/admin/AdminPage";
-import { CollectionDetailPage } from "./components/admin/CollectionDetailPage";
-import { TableDetailPage } from "./components/admin/TableDetailPage";
 import "./styles/global.css";
 
 function AppContent() {
@@ -128,9 +125,6 @@ function AppContent() {
           <Route path="/lexicons" element={<LexiconView />} />
           <Route path="/docs" element={<DocsPage />} />
           <Route path="/settings" element={<SettingsPage />} />
-          <Route path="/admin" element={<AdminPage />} />
-          <Route path="/admin/collections/:nsid" element={<CollectionDetailPage />} />
-          <Route path="/admin/tables/:name" element={<TableDetailPage />} />
           <Route path="*" element={<NotFound />} />
         </Routes>
       </Box>
